Split admin insights chart surfaces
